chore(admin-server): move variables to outside handler
