+++ packages/admin-server/src/app.ts
@@ -1,51 +1,51 @@
-import Koa, {Context} from 'koa';
+import Koa from 'koa';
 import path from 'path';
 import { Issuer } from 'openid-client';
 import views from 'koa-views';
 import serve from 'koa-static';
 import Router from 'koa-router';
 import morgan from 'koa-morgan';
-
 import Agent, { HttpsAgent } from 'agentkeepalive';
-import koaMount from 'koa-mount';
 
 // controller
 import { OidcCtrl, mount as mountOidc } from './oidc';
-import { ProxyCtrl} from './proxy';
+import { ProxyCtrl } from './proxy';
 
 // config
-import {createConfig, Config} from './config';
+import { createConfig, Config } from './config';
+import createEnvVariablesSetup from './createEnvVariablesSetup';
+import errorHandler from './errorHandler';
 
 // logger
 import * as logger from './logger';
 
-export const createApp = async (): Promise<{app: Koa, config: Config}> => {
+export const createApp = async (): Promise<{ app: Koa; config: Config }> => {
   const config = createConfig();
   const staticPath = config.appPrefix ? `${config.appPrefix}/` : '/';
 
   // construct http agent
   const httpAgent = new Agent({
     maxSockets: config.keycloakMaxSockets,
-    maxFreeSockets: config.keycloakMaxFreeSockets
+    maxFreeSockets: config.keycloakMaxFreeSockets,
   });
 
   const httpsAgent = new HttpsAgent({
     maxSockets: config.keycloakMaxSockets,
-    maxFreeSockets: config.keycloakMaxFreeSockets
+    maxFreeSockets: config.keycloakMaxFreeSockets,
   });
 
   // create oidc client and controller
   Issuer.defaultHttpOptions = {
     agent: {
       http: httpAgent,
-      https: httpsAgent
+      https: httpsAgent,
     },
     retries: config.keycloakRetries,
     timeout: config.keycloakTimeout,
   };
 
   // tslint:disable-next-line:max-line-length
-  const issuer =  new Issuer({
+  const issuer = new Issuer({
     issuer: `${config.keycloakOidcBaseUrl}/realms/${config.keycloakRealmName}`,
     authorization_endpoint: `${config.keycloakOidcBaseUrl}/realms/${config.keycloakRealmName}/protocol/openid-connect/auth`,
     token_endpoint: `${config.keycloakApiBaseUrl}/realms/${config.keycloakRealmName}/protocol/openid-connect/token`,
@@ -55,7 +55,7 @@ export const createApp = async (): Promise<{app: Koa, config: Config}> => {
 
   const oidcClient = new issuer.Client({
     client_id: config.keycloakClientId,
-    client_secret: config.keycloakClientSecret
+    client_secret: config.keycloakClientSecret,
   });
   oidcClient.CLOCK_TOLERANCE = 5 * 60;
   const oidcCtrl = new OidcCtrl({
@@ -64,73 +64,11 @@ export const createApp = async (): Promise<{app: Koa, config: Config}> => {
   });
 
   // koa
-  const app = new Koa() as any;
+  const app = new Koa();
   app.keys = [config.cookieSignedKey];
 
   // set proxy=true to support https behind proxy
   app.proxy = true;
-
-  // setup
-  app.use(async (ctx: Context, next) => {
-    ctx.state.locale = config.locale;
-    ctx.state.cmsHost = config.cmsHost;
-    ctx.state.graphqlPrefix = config.graphqlPrefix;
-    ctx.state.graphqlEndpoint = config.graphqlEndpoint;
-    ctx.state.requestApiTokenEndpoint = config.appPrefix ? `${config.appPrefix}/oidc/request-api-token` : '/oidc/request-api-token';
-    ctx.state.disableMode = config.readOnlyOnInstanceTypeAndImage;
-    ctx.state.enableDatasetUpload = config.enableDatasetUpload;
-    ctx.state.enableCustomImage = config.enableCustomImage;
-    ctx.state.enableMaintenanceNotebook = config.enableMaintenanceNotebook;
-    ctx.state.enableGrafana = config.enableGrafana;
-    ctx.state.enableUsageReport = config.enableUsageReport;
-    ctx.state.customImageSetup = config.customImageSetup;
-    ctx.state.enableLicenseCheck = config.enableLicenseCheck;
-    ctx.state.enableModelDeployment = config.enableModelDeployment;
-    ctx.state.enableLogPersistence = config.enableLogPersistence;
-    ctx.state.enablePhfs = config.enablePhfs;
-    ctx.state.enableJobArtifact = config.enableJobArtifact;
-    ctx.state.enableJobMonitoring = config.enableJobMonitoring;
-    ctx.state.enableApp = config.enableApp;
-    ctx.state.disableGroup = config.enableLicenseCheck ? config.licenseStatus !== 'unexpired' : false;
-    ctx.state.everyoneGroupId = config.keycloakEveryoneGroupId;
-    ctx.state.jobDefaultActiveDeadlineSeconds = config.jobDefaultActiveDeadlineSeconds;
-    ctx.state.primehubVersion = config.primehubVersion;
-    ctx.state.isUserAdmin = config.isUserAdmin;
-    ctx.state.newAccessToken = config.newAccessToken;
-
-    // referrer
-    const referrer = `${config.cmsHost}${ctx.path}`;
-    ctx.state.links = JSON.stringify({
-      // tslint:disable-next-line:max-line-length
-      userProfileLink: `${config.keycloakOidcBaseUrl}/realms/${config.keycloakRealmName}/account?referrer=${config.keycloakClientId}&referrer_uri=${referrer}`,
-      // tslint:disable-next-line:max-line-length
-      changePasswordLink: `${config.keycloakOidcBaseUrl}/realms/${config.keycloakRealmName}/account/password?referrer=${config.keycloakClientId}&referrer_uri=${referrer}`,
-      adminPortalLink: config.appPrefix ? `${config.appPrefix}/cms` : '/cms',
-      logoutLink: config.appPrefix ? `${config.appPrefix}/oidc/logout` : '/oidc/logout',
-    });
-
-    return next();
-  });
-
-  // error handler
-  app.use(async (ctx: Context, next) => {
-    try {
-      await next();
-    } catch (err) {
-      const errorCode = (err.isBoom && err.data && err.data.code) ? err.data.code : 'INTERNAL_ERROR';
-      const statusCode =
-        (err.isBoom && err.output && err.output.statusCode) ? err.output.statusCode : err.status || 500;
-
-      ctx.status = statusCode;
-
-      // render or json
-      if (ctx.accepts('html') && ctx.status === 403) {
-        return ctx.render('403', {message: err.message, staticPath});
-      } else {
-        ctx.body = {code: errorCode, message: err.message};
-      }
-    }
-  });
 
   if (!process.env.TEST) {
     const morganFormat: any = (tokens, req, res) => {
@@ -145,21 +83,27 @@ export const createApp = async (): Promise<{app: Koa, config: Config}> => {
         status: parseInt(tokens.status(req, res), 10),
         contentLength: tokens.res(req, res, 'content-length'),
         responseTime: `${tokens['response-time'](req, res)} ms`,
-        userAgent: req.headers['user-agent']
+        userAgent: req.headers['user-agent'],
       });
     };
     app.use(morgan(morganFormat));
   }
 
-  app.use(views(path.resolve(__dirname, '../../client/dist'), { extension: 'ejs' }));
-  const serveClientStatic = config.appPrefix
-    ? koaMount(config.appPrefix, serve(path.resolve(__dirname, '../../client/dist'), {gzip: true, index: false}))
-    : serve(path.resolve(__dirname, '../../client/dist'), {gzip: true, index: false});
-  app.use(serveClientStatic);
+  app.use(createEnvVariablesSetup());
+  app.use(errorHandler());
+  app.use(
+    views(path.resolve(__dirname, '../../client/dist'), { extension: 'ejs' })
+  );
+  app.use(
+    serve(path.resolve(__dirname, '../../client/dist'), {
+      gzip: true,
+      index: false,
+    })
+  );
 
   // router
   const rootRouter = new Router({
-    prefix: config.appPrefix
+    prefix: config.appPrefix,
   });
 
   // redirect
@@ -216,6 +160,8 @@ export const createApp = async (): Promise<{app: Koa, config: Config}> => {
     oidcCtrl,
   });
   proxy.mount(rootRouter);
+
+  // @ts-ignore
   app.upgradeHandler = proxy.createUpgradeHandler();
 
   // health check
@@ -223,5 +169,6 @@ export const createApp = async (): Promise<{app: Koa, config: Config}> => {
     ctx.status = 200;
   });
   app.use(rootRouter.routes());
-  return {app, config};
+
+  return { app, config };
 };
